rename pack dir name
