+++ Packs/Twitter/Integrations/Twitter/Twitter.yml
@@ -1,0 +1,96 @@
+commonfields:
+  id: Twitter
+  version: -1
+name: Twitter
+display: Twitter
+category: Utilities
+description: 'The Twitter Integration allows users to parse Twitter for Users,
+  Tweets, and additional info about users. Perform enhanced searches with additional
+  search arguments. Search results are returned as a markdown table. '
+configuration:
+- display: API Key
+  name: apikey
+  type: 4
+  required: true
+- display: API Key Secret
+  name: apikey_secret
+  type: 4
+  required: true
+- display: Access Token
+  name: access_token
+  type: 4
+  required: true
+- display: Access Token Secret
+  name: access_token_secret
+  type: 4
+  required: true
+- display: Bearer Token
+  name: bearer_token
+  type: 4
+  required: true
+script:
+  script: '-'
+  type: python
+  commands:
+  - name: twitter-get-users
+    arguments:
+    - name: name
+      required: true
+      description: The string to search for accounts with similar names
+    - name: count
+      description: 'Number of accounts to return. Note: The max is 20. Any number
+        greater than 20 will return 20 accounts.'
+      defaultValue: "15"
+    - name: page
+      description: Page to look for search results on
+      defaultValue: "1"
+    outputs:
+    - contextPath: TwitterSOARx
+      description: Results of the twitter-get-users command
+      type: string
+    description: Search for public users on Twitter
+  - name: twitter-get-user-info
+    arguments:
+    - name: usernames
+      description: 'The twitter usernames to search'
+      required: true
+    outputs:
+    - contextPath: TwitterSOARx
+      description: Results of the twitter-get-user-info command
+      type: string
+    description: 'Lookup users by name to display information about them. Search multiple
+      users simultaneously by separating them by commas. Ex: "name="user1,user2,user3"'
+  - name: twitter-get-tweets
+    arguments:
+    - name: q
+      required: true
+      description: The word or string of words to search for in tweets.
+    - name: result_type
+      description: Recent, Popular, Mixed
+    - name: geocode
+      description: 'GPS Coordinates: Latitude,Longitude,RadiusMi/KM. Ex: geocode="23.453987,-35.726374,100mi"'
+    - name: lang
+      description: 'The language code of tweets to search for. List of supported languages:  https://developer.twitter.com/en/docs/twitter-for-websites/supported-languages'
+    - name: count
+      description: The number of tweets to return. The max is 100. Any number greater
+        than 100  will return 100.
+    - name: include_entities
+      description: Display entities (True/False), Defaults to True
+      defaultValue: "True"
+    - name: from_user
+      description: 'Search tweets from a given user. Ex: "from_user=PaloAltoNtwks"
+        returns only tweets from Palo Alto Networks'
+    - name: to_user
+      description: 'Search tweets to a given user. Ex: "to_user=PaloAltoNtwks" returns
+        only tweets in reply to Palo Alto Networks'
+    outputs:
+    - contextPath: TwitterSOARx
+      description: Results of the twitter-get-tweets command
+      type: string
+    description: Search for tweets on Twitter.
+  dockerimage: demisto/tweepy:1.0.0.23810
+  runonce: false
+  subtype: python3
+fromversion: 6.0.0
+tests:
+- No tests (auto formatted)
